fix option parsing, add missing help (#45)
In light of the discussion in #44, option parsing needs to be more
strict. We shouldn't allow an orphaned config at the end of a line that
looks like an option to another argument. This commit ensures that the
config is either passed FIRST, or via the -c/--config option. It also
does full argument validation before calling showStatus or showHelp.